MultimodalLeanRAG/requirements.txt
# Multimodal LeanRAG Requirements
# ================================

# Core dependencies (inherited from LeanRAG)
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Vector Database
pymilvus>=2.3.0
milvus-lite>=2.4.0

# LLM and Embeddings
openai>=1.0.0
tiktoken>=0.5.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
transformers>=4.35.0
torch>=2.0.0
torchaudio>=2.0.0

# CLAP Audio Embeddings
laion-clap>=1.1.4

# Graph and Database
networkx>=3.0
pymysql>=1.1.0

# Utilities
joblib>=1.3.0
requests>=2.31.0
hashlib-compat>=1.0.0

# Clustering
umap-learn>=0.5.0
hdbscan>=0.8.0

# Optional: For visual modality
# Pillow>=10.0.0
# opencv-python>=4.8.0
# open-clip-torch>=2.20.0
